Move str2dict conversion out of getdns_query
